Remove utils/comments file
