+++ github-actions/issue-labels/post-labels-comment.js
@@ -1,4 +1,4 @@
-const comments = require("../utils/comments")
+var fs = require("fs")
 
 // Constant variables
 const LABELS_OBJ = {
@@ -35,9 +35,9 @@ async function main({ g, c }, { actionResult, addedLabels, issueNum }) {
   const issueCreatorPlaceholder = '${issueCreator}'
 
   const instructions = makeComment(addedLabels)
-  const formattedInstructions = comments.formatComment(instructions, instructionsPlaceholder, path, null)
-  const instructionsWithIssueCreator = comments.formatComment(issueCreator, issueCreatorPlaceholder, null, formattedInstructions)
-  await comments.postComment(github, context, issueNum, instructionsWithIssueCreator)
+  const formattedInstructions = formatComment(instructions, instructionsPlaceholder, path, null)
+  const instructionsWithIssueCreator = formatComment(issueCreator, issueCreatorPlaceholder, null, formattedInstructions)
+  await postComment(github, context, issueNum, instructionsWithIssueCreator)
 }
 
 /**
@@ -48,13 +48,47 @@ async function main({ g, c }, { actionResult, addedLabels, issueNum }) {
 function makeComment(labels) {
   if (labels.length === 0) {
     const path = './github-actions/issue-labels/no-labels-template.md'
-    return comments.formatComment(null, null, path, null)
+    return formatComment(null, null, path, null)
   }
 
   const path = './github-actions/issue-labels/add-labels-template.md'
   const labelsPlaceholder = '${labels}'
   const labelsToAdd = labels.map(label => LABELS_OBJ[label]).join(', ')
-  return comments.formatComment(labelsToAdd, labelsPlaceholder, path, null)
+  return formatComment(labelsToAdd, labelsPlaceholder, path, null)
+}
+
+/**
+ * Formats the comment to be posted
+ * @param {String} replacementString - the string to replace the placeholder in the md file
+ * @param {String} placeholderString - the placeholder to be replaced in the md file
+ * @param {String} path - the path of the md file to be formatted
+ * @param {String} textToFormat - the text to be formatted. If null, use the md file provided in the path. If provided, format that text
+ * @returns {String} - returns a formatted comment to be posted on github
+ */
+ function formatComment(replacementString, placeholderString, path, textToFormat) {
+  const text = textToFormat === null ? fs.readFileSync(path).toString('utf-8') : textToFormat
+  const commentToPost = text.replace(placeholderString, replacementString)
+  return commentToPost
+}
+
+/**
+ * Posts a comment on github
+ * @param {Object} github - github object  
+ * @param {Object} context - context object 
+ * @param {Number} issueNum - the issue number where the comment should be posted
+ * @param {String} comment - the comment to be posted
+ */
+ async function postComment(github, context, issueNum, comment) {
+  try {
+    await github.issues.createComment({
+      owner: context.repo.owner,
+      repo: context.repo.repo,
+      issue_number: issueNum,
+      body: comment,
+    })
+  } catch (err) {
+    throw new Error(err);
+  }
 }
 
 module.exports = main
